requirements.txt
flask
waitress
torch
torchvision
opencv-python
numpy
pillow
scikit-learn
xgboost
efficientnet-pytorch